requirements.txt
python-dotenv
llama-index
llama-index-vector-stores-chroma
llama-index-llms-anthropic
llama-index-llms-cohere
llama-index-postprocessor-cohere-rerank
llama-parse
tiktoken
chromadb
gradio
tqdm
cohere
lightrag-hku
